# docs/verification.openapi.yml
openapi: 3.0.0
info:
  title: MicroMerit Portal - Verification API
  version: 1.0.0
  description: |
    API for verifying credentials and PDF certificates in the MicroMerit Portal.
    These are public endpoints that allow anyone to verify the authenticity of credentials.

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.micromerit.com
    description: Production server

tags:
  - name: Verification
    description: Credential and PDF verification endpoints (public access)

paths:
  /verify/{credential_uid}:
    get:
      tags:
        - Verification
      summary: Verify credential by UID
      description: |
        Public endpoint to verify a credential by its unique identifier.
        Returns complete credential information including issuer details, learner info (if claimed),
        PDF URLs, metadata, and blockchain record (or placeholder if pending).
      operationId: verifyCredentialByUid
      parameters:
        - name: credential_uid
          in: path
          required: true
          description: Unique identifier of the credential
          schema:
            type: string
            example: CRED-1731888000000-A1B2C3D4E5F6G7H8
      responses:
        '200':
          description: Credential verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Credential verified successfully
                  data:
                    $ref: '#/components/schemas/VerifiedCredential'
              examples:
                claimedCredential:
                  summary: Claimed credential with blockchain record
                  value:
                    success: true
                    message: Credential verified successfully
                    data:
                      success: true
                      credential:
                        uid: CRED-1731888000000-A1B2C3D4E5F6G7H8
                        status: claimed
                        issuedAt: '2024-01-15T10:30:00.000Z'
                        claimedAt: '2024-01-16T14:20:00.000Z'
                        metadata:
                          title: Certificate of Completion
                          description: Successfully completed the Advanced Web Development course
                          courseCode: CS301
                          grade: A
                          credits: 4
                      issuer:
                        id: 1
                        name: Tech University
                        type: university
                        logoUrl: https://s3.example.com/logos/tech-uni.png
                        officialDomain: tech.edu
                        websiteUrl: https://tech.edu
                        email: credentials@tech.edu
                        status: approved
                      learner:
                        id: 5
                        email: student@example.com
                        phone: '+1234567890'
                        profileUrl: https://s3.example.com/profiles/student.jpg
                      pdf:
                        pdfUrl: https://s3.example.com/certificates/CRED-123-certificate.pdf
                        qrCodeUrl: https://s3.example.com/certificates/CRED-123-qr.png
                        createdAt: '2024-01-15T10:35:00.000Z'
                      blockchain:
                        transactionId: '0xabc123def456...'
                        hashValue: 'a1b2c3d4e5f6...'
                        storedAt: '2024-01-15T10:40:00.000Z'
                unclaimedCredential:
                  summary: Unclaimed credential with pending blockchain
                  value:
                    success: true
                    message: Credential verified successfully
                    data:
                      success: true
                      credential:
                        uid: CRED-1731888000000-B2C3D4E5F6G7H8I9
                        status: issued
                        issuedAt: '2024-02-01T09:00:00.000Z'
                        claimedAt: null
                        metadata:
                          title: Professional Certification
                          description: Certified in Project Management
                      issuer:
                        id: 2
                        name: Professional Certifications Inc
                        type: training_provider
                        logoUrl: null
                        officialDomain: profcert.com
                        websiteUrl: https://profcert.com
                        email: info@profcert.com
                        status: approved
                      learner: null
                      pdf: null
                      blockchain:
                        transactionId: pending
                        hashValue: pending
                        storedAt: null
                        note: Blockchain verification pending
        '404':
          description: Credential not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Credential not found
                error: NOT_FOUND
                statusCode: 404

  /verify/pdf:
    post:
      tags:
        - Verification
      summary: Verify PDF certificate upload
      description: |
        Public endpoint to verify a PDF certificate by uploading the file.
        
        **Verification Process:**
        1. Extracts credential UID from PDF text
        2. Fetches credential from database
        3. Computes SHA256 hash of uploaded PDF
        4. Compares with stored hash in metadata
        5. Checks credential status (revoked/issued/claimed)
        
        **Hash Verification:**
        - If hash matches: PDF is authentic and unmodified
        - If hash mismatches: PDF has been tampered with
        - If no stored hash: Can verify existence but not integrity
        
        **Credential UID Detection:**
        The system looks for patterns like:
        - "Credential ID: CRED-..."
        - "CRED-1234567890-ABCDEF"
        - "ID: CRED-..."
      operationId: verifyPdfUpload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - pdf_file
              properties:
                pdf_file:
                  type: string
                  format: binary
                  description: PDF certificate file to verify (max 10MB)
      responses:
        '200':
          description: PDF verification completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: PDF verification completed
                  data:
                    $ref: '#/components/schemas/PdfVerificationResult'
              examples:
                verifiedSuccess:
                  summary: PDF verified successfully
                  value:
                    success: true
                    message: PDF verification completed
                    data:
                      verified: true
                      reason: PDF verification successful - authentic and unmodified
                      hashMatch: true
                      credential:
                        uid: CRED-1731888000000-A1B2C3D4E5F6G7H8
                        status: claimed
                        issuedAt: '2024-01-15T10:30:00.000Z'
                        claimedAt: '2024-01-16T14:20:00.000Z'
                        metadata:
                          title: Certificate of Completion
                      issuer:
                        id: 1
                        name: Tech University
                        type: university
                        logoUrl: https://s3.example.com/logos/tech-uni.png
                        officialDomain: tech.edu
                        websiteUrl: https://tech.edu
                      learner:
                        id: 5
                        email: student@example.com
                        phone: '+1234567890'
                        profileUrl: https://s3.example.com/profiles/student.jpg
                      blockchain:
                        transactionId: '0xabc123def456...'
                        hashValue: 'a1b2c3d4e5f6...'
                        storedAt: '2024-01-15T10:40:00.000Z'
                tamperedPdf:
                  summary: PDF has been tampered with
                  value:
                    success: true
                    message: PDF verification completed
                    data:
                      verified: false
                      reason: PDF has been tampered with (hash mismatch)
                      hashMatch: false
                      credential:
                        uid: CRED-1731888000000-A1B2C3D4E5F6G7H8
                        status: issued
                        issuedAt: '2024-01-15T10:30:00.000Z'
                      issuer:
                        id: 1
                        name: Tech University
                      learner: null
                noCredentialUid:
                  summary: No credential UID found in PDF
                  value:
                    success: true
                    message: PDF verification completed
                    data:
                      verified: false
                      reason: No credential UID found in PDF
                      credential: null
                      issuer: null
                      learner: null
                revokedCredential:
                  summary: Credential has been revoked
                  value:
                    success: true
                    message: PDF verification completed
                    data:
                      verified: false
                      reason: Credential has been revoked by issuer
                      hashMatch: true
                      credential:
                        uid: CRED-1731888000000-REVOKED123
                        status: revoked
                        issuedAt: '2024-01-15T10:30:00.000Z'
                        claimedAt: '2024-01-16T14:20:00.000Z'
                      issuer:
                        id: 1
                        name: Tech University
                        type: university
                      learner:
                        id: 5
                        email: student@example.com
                noStoredHash:
                  summary: No stored hash available
                  value:
                    success: true
                    message: PDF verification completed
                    data:
                      verified: true
                      reason: Credential exists but PDF hash verification not available (no stored hash)
                      hashMatch: null
                      credential:
                        uid: CRED-1731888000000-NOHASH456
                        status: issued
                        issuedAt: '2024-01-15T10:30:00.000Z'
                        claimedAt: null
                        metadata:
                          title: Old Certificate
                      issuer:
                        id: 1
                        name: Tech University
                        type: university
                        logoUrl: https://s3.example.com/logos/tech-uni.png
                      learner: null
        '400':
          description: Bad request (invalid file, file too large, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                noFile:
                  summary: No file uploaded
                  value:
                    success: false
                    message: No PDF file uploaded. Upload a PDF file using the field name "pdf_file"
                    error: NO_FILE
                    statusCode: 400
                invalidFileType:
                  summary: Invalid file type
                  value:
                    success: false
                    message: Invalid file type image/png. Expected application/pdf
                    error: INVALID_FILE_TYPE
                    statusCode: 400
                fileTooLarge:
                  summary: File too large
                  value:
                    success: false
                    message: File size 12582912 exceeds maximum 10485760 bytes (10MB)
                    error: FILE_TOO_LARGE
                    statusCode: 400
                corruptedPdf:
                  summary: Corrupted PDF file
                  value:
                    success: false
                    message: Invalid or corrupted PDF file. Please ensure you are uploading a valid PDF certificate
                    error: INVALID_PDF
                    statusCode: 400

components:
  schemas:
    VerifiedCredential:
      type: object
      properties:
        success:
          type: boolean
          example: true
        credential:
          type: object
          properties:
            uid:
              type: string
              example: CRED-1731888000000-A1B2C3D4E5F6G7H8
            status:
              type: string
              enum: [issued, claimed, revoked]
              example: claimed
            issuedAt:
              type: string
              format: date-time
              example: '2024-01-15T10:30:00.000Z'
            claimedAt:
              type: string
              format: date-time
              nullable: true
              example: '2024-01-16T14:20:00.000Z'
            metadata:
              type: object
              description: Custom metadata set by issuer
              additionalProperties: true
        issuer:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Tech University
            type:
              type: string
              enum: [university, edtech, company, training_provider]
              example: university
            logoUrl:
              type: string
              nullable: true
              example: https://s3.example.com/logos/tech-uni.png
            officialDomain:
              type: string
              nullable: true
              example: tech.edu
            websiteUrl:
              type: string
              nullable: true
              example: https://tech.edu
            email:
              type: string
              example: credentials@tech.edu
            status:
              type: string
              enum: [pending, approved, rejected]
              example: approved
        learner:
          type: object
          nullable: true
          description: Learner info if credential is claimed, null if unclaimed
          properties:
            id:
              type: integer
              example: 5
            email:
              type: string
              nullable: true
              example: student@example.com
            phone:
              type: string
              nullable: true
              example: '+1234567890'
            profileUrl:
              type: string
              nullable: true
              example: https://s3.example.com/profiles/student.jpg
        pdf:
          type: object
          nullable: true
          description: PDF certificate info if generated, null if not
          properties:
            pdfUrl:
              type: string
              example: https://s3.example.com/certificates/CRED-123-certificate.pdf
            qrCodeUrl:
              type: string
              nullable: true
              example: https://s3.example.com/certificates/CRED-123-qr.png
            createdAt:
              type: string
              format: date-time
              example: '2024-01-15T10:35:00.000Z'
        blockchain:
          type: object
          description: Blockchain record or placeholder if pending
          properties:
            transactionId:
              type: string
              example: '0xabc123def456...'
            hashValue:
              type: string
              example: 'a1b2c3d4e5f6...'
            storedAt:
              type: string
              format: date-time
              nullable: true
              example: '2024-01-15T10:40:00.000Z'
            note:
              type: string
              example: Blockchain verification pending
              description: Present when blockchain record is pending

    PdfVerificationResult:
      type: object
      properties:
        verified:
          type: boolean
          description: Whether the PDF is verified as authentic
          example: true
        reason:
          type: string
          description: Explanation of verification result
          example: PDF verification successful - authentic and unmodified
        hashMatch:
          type: boolean
          nullable: true
          description: Whether PDF hash matches stored hash (null if no stored hash)
          example: true
        credential:
          type: object
          nullable: true
          description: Credential details if found
          properties:
            uid:
              type: string
            status:
              type: string
            issuedAt:
              type: string
              format: date-time
            claimedAt:
              type: string
              format: date-time
              nullable: true
            metadata:
              type: object
              additionalProperties: true
        issuer:
          type: object
          nullable: true
          description: Issuer details if credential found
          properties:
            id:
              type: integer
            name:
              type: string
            type:
              type: string
            logoUrl:
              type: string
              nullable: true
            officialDomain:
              type: string
              nullable: true
            websiteUrl:
              type: string
              nullable: true
        learner:
          type: object
          nullable: true
          description: Learner details if credential is claimed
          properties:
            id:
              type: integer
            email:
              type: string
              nullable: true
            phone:
              type: string
              nullable: true
            profileUrl:
              type: string
              nullable: true
        blockchain:
          type: object
          nullable: true
          description: Blockchain record if available
          properties:
            transactionId:
              type: string
            hashValue:
              type: string
            storedAt:
              type: string
              format: date-time

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: ERROR_CODE
        statusCode:
          type: integer
          example: 400